Added the Best First Search at this point
- Added the bestFS() function.
- Changed the name of the stack from dfsStack to searchStack hoping to
make it usable for more search types.
- Added an ArrayList that holds the Stack of TileTrees different for
different scores. So, lets say we have 5 TileTrees in the board where we
can get score of 2, the these TileTrees will be stored in a Stack in
index 2.
- This helped in pushing the TileTrees in open stack according to the
order of score. (Taking first highest score).
- Made a path stack which will hold the optimal end state path.
CONCEPT: Take a new tile -- Find the available locations of it in the
board -- If available, sort it according to the score and get the first
location with highest score.  -- Push it to the path -- Print it --  If
no new location is found, then put the new tile back to stock and make
the new tile as popped tile from open tile tree stack.